[김경덕 남상현] 채팅창 스크롤 추가 Qna sidemenu 추가 Qna session검사 Notice sidemenu 추가 Notice session검사

+++ secondhand/src/main/java/com/project/secondhand/controller/NoticeController.java
@@ -21,13 +21,9 @@ public class NoticeController {
 	//입력하기 form(운영자)
 	@GetMapping("/addNotice")
 	public String addNotice(HttpSession session) {
-		
-			
-//		if(관리자가 아닐때) {
-//			return "redirect:/index";
-//		}
-		
-		// notice 입력폼 html
+		if(session.getAttribute("loginAdmin") == null) {
+			return "redirect:/";
+		}
 		return "/addNotice";
 	}
 	
@@ -35,10 +31,9 @@ public class NoticeController {
 	@PostMapping("/addNotice")
 	public String addNotice(HttpSession session, Notice notice) {
 		System.out.println(notice.toString());
-//		if(관리자가 아닐때) {
-//			return "redirect:/index";
-//		} 
-		
+		if(session.getAttribute("loginAdmin") == null) {
+			return "redirect:/";
+		}
 		noticeService.addNotice(notice);
 		// notice list로
 		return "redirect:/noticeList";
@@ -59,7 +54,6 @@ public class NoticeController {
 	// 공지사항 상세보기
 	@GetMapping("/noticeDetail")
 	public String getNotice(@RequestParam(value="noticeNo") int noticeNo, Model model) {
-		System.out.println("----------------------------------");
 		model.addAttribute("notice", noticeService.getNoticeOne(noticeNo));
 		// notice detail 페이지
 		return "noticeDetail";
@@ -69,11 +63,9 @@ public class NoticeController {
 	// 공지사항 수정 form(운영자)
 	@GetMapping("/modifyNotice")
 	public String modifyNotice(HttpSession session , @RequestParam(value="noticeNo") int noticeNo, Model model) {
-//		
-//		if(관리자가 아닐때) {
-//			return "redirect:/index";
-//		}
-		
+		if(session.getAttribute("loginAdmin") == null) {
+			return "redirect:/";
+		}
 		model.addAttribute("notice", noticeService.getNoticeOne(noticeNo));
 		// notice 수정폼 html
 		return "modifyNotice";
@@ -82,25 +74,20 @@ public class NoticeController {
 	// 공지사항 수정 action(운영자)
 	@PostMapping("/modifyNotice")
 	public String modifyNotice(HttpSession session, Notice notice) {
-		System.out.println("<---------------------------------------");
-//		if(관리자가 아닐때) {
-//			return "redirect:/index";
-//		}
-//		
+		if(session.getAttribute("loginAdmin") == null) {
+			return "redirect:/";
+		}
 		noticeService.modifyNotice(notice);
 		// 수정한 게시물 페이지로
-		// ex) return "redirect:/상세보기주소?noticeNo=notice.getNoticeNo";
 		return "redirect:/noticeDetail?noticeNo="+ notice.getNoticeNo();
 	}
 	
 	// 공지사항 삭제(운영자)
 	@GetMapping("/removeNotice")
 	public String removeNotice(HttpSession session, @RequestParam(value="noticeNo") int noticeNo) {
-		
-//		if(관리자가 아닐때) {
-//			return "redirect:/index";
-//		}
-		
+		if(session.getAttribute("loginAdmin") == null) {
+			return "redirect:/";
+		}
 		noticeService.removeNotice(noticeNo);
 		// notice list 페이지로
 		return "redirect:/noticeList";
