+++ secondhand/src/main/java/com/project/secondhand/controller/QnaController.java
@@ -51,10 +51,9 @@ public class QnaController {
 	@GetMapping("/addQna")
 	public String addQna(HttpSession session, Model model) {
 		System.out.println("/addQna 요청하기");
-//			관리자로그인
-//		if(관리자가 아닐떄) {
-//			return "redirect:"/"
-//		}
+		if(session.getAttribute("loginAdmin") == null) {
+			return "redirect:/";
+		}
 		List<Category> categoryList = categoryService.getCategoryList("qna");
 		System.out.println(categoryList + "<--------categoryList");
 		model.addAttribute("categoryList", categoryList);
@@ -64,10 +63,9 @@ public class QnaController {
 	//자주묻는 질문입력하기 action(운영자)
 	@PostMapping("/addQna")
 	public String addQna(HttpSession session, Qna qna) {
-//			관리자 로그인
-//		if(관리자가 아닐떄) {
-//			return "redirect:"/"
-//		}
+		if(session.getAttribute("loginAdmin") == null) {
+			return "redirect:/";
+		}
 		qnaService.addQna(qna);
 		return "redirect:/qnaList";
 	}
@@ -75,10 +73,6 @@ public class QnaController {
 	//자주 묻는 질문 리스트
 	@GetMapping("/qnaList")
 	public String getQnaList(HttpSession session, Model model, @RequestParam(value = "searchWord",required = false)String searchWord) {
-//		관리자 로그인
-//	if(관리자가 아닐떄) {
-//		return "redirect:"/"
-//	}
 		if(searchWord==null) {
 			searchWord="";
 		}
@@ -91,10 +85,6 @@ public class QnaController {
 	//자주 묻는 질문 상세보기
 	@GetMapping("/qnaDetail")
 	public String getDetail(HttpSession session, Model model, @RequestParam(value="qnaTitle")String qnaTitle) {
-//		관리자 로그인							
-//	if(관리자가 아닐떄) {
-//		return "redirect:"/"
-//	}
 		System.out.println(qnaTitle +"<-------------qnatitle qnaDetail");
 		model.addAttribute("qna", qnaService.getQna(qnaTitle));
 		return "qnaDetail";
@@ -102,10 +92,9 @@ public class QnaController {
 //	자주 묻는 질문 수정하기 form(운영자)
 	@GetMapping("/modifyQna")
 	public String modifyQna(HttpSession session, Model model, @RequestParam(value ="qnaTitle")String qnaTitle) {
-//		관리자 로그인							
-//	if(관리자가 아닐떄) {
-//		return "redirect:"/"
-//	}
+		if(session.getAttribute("loginAdmin") == null) {
+			return "redirect:/";
+		}
 		List<Category> list = categoryService.getCategoryList("qna");	
 		model.addAttribute("qna", qnaService.getQna(qnaTitle));
 		model.addAttribute("categoryList", list);
@@ -114,10 +103,9 @@ public class QnaController {
 	//자주 묻는 질문 수정하기 action(운영자)
 	@PostMapping("/modifyQna")
 	public String modifyQna(HttpSession session, Qna qna, @RequestParam(value="originTitle") String originTitle) {
-//		관리자 로그인							
-//	if(관리자가 아닐떄) {
-//		return "redirect:"/"
-//	}
+		if(session.getAttribute("loginAdmin") == null) {
+			return "redirect:/";
+		}
 		if(originTitle.equals(qna.getQnaTitle())) {
 			qnaService.modifyQna(qna);
 		} else {
@@ -135,10 +123,9 @@ public class QnaController {
 	//자주 묻는 질문 삭제하기
 	@GetMapping("/removeQna")
 	public String removeQna(HttpSession session, @RequestParam(value ="qnaTitle", required = false)String qnaTitle) {
-//		관리자 로그인							
-//	if(관리자가 아닐떄) {
-//		return "redirect:"/"
-//	}
+		if(session.getAttribute("loginAdmin") == null) {
+			return "redirect:/";
+		}
 		qnaService.removeQna(qnaTitle);
 		return "redirect:/qnaList";
 	}  
@@ -146,10 +133,9 @@ public class QnaController {
 	//메일 문의하기(Form)
 	@GetMapping("/sendQna")
 	public String sendQna(HttpSession session, Model model) {
-//									
-//	if(로그인이 안되어있을때) {
-//		return "redirect:"/"
-//	}
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/";
+		}
 		List<Category> list = categoryService.getCategoryList("qna");
 		model.addAttribute("categoryList", list);
 		return "sendQna";
@@ -157,14 +143,12 @@ public class QnaController {
 	//메일 문의하기(Action)
 	@PostMapping("/sendQna")
 	public String sendQna(HttpSession session, @RequestParam(value = "sendQnaCategory")String sendQnaCategory, @RequestParam(value = "sendQnaContent")String sendQnaContent) {
-//		
-//if(로그인이 안되어있을때) {
-//return "redirect:"/"
-//}	
+		if(session.getAttribute("loginMember") == null) {
+			return "redirect:/";
+		}
 		System.out.println(sendQnaCategory + "<-------------sendQnaCategory");
 		System.out.println(sendQnaContent + "<-------------sendQnaContent");
-		//String memberEmail = ((Member)session.getAttribute("loginMember")).getMemberEmail();
-		String memberEmail = "test";
+		String memberEmail = ((Member)session.getAttribute("loginMember")).getMemberEmail();
 		qnaService.sendQna(memberEmail, sendQnaCategory, sendQnaContent);
 		return "redirect:/qnaList";
 	}
